Undo changes in infrastucture and return miss_rate_sim
