+++ simulator/infra/cache/t/miss_rate.csv
@@ -1,0 +1,6 @@
+0.349048, 0.27356, 0.216863, 0.164309, 0.127481, 0.0896536, 0.0614431, 0.0385261, 0.0342199, 0.0278209, 0.0277584, 
+0.320341, 0.23851, 0.179896, 0.143943, 0.110717, 0.0795007, 0.0479722, 0.0332808, 0.027367, 0.0273402, 0.0271651, 
+0.306841, 0.229941, 0.168384, 0.133186, 0.106952, 0.0773909, 0.043062, 0.0287753, 0.0273411, 0.0271651, 0.0271651, 
+0.303847, 0.226837, 0.166197, 0.130631, 0.104856, 0.0762453, 0.0409576, 0.0273447, 0.0271651, 0.0271651, 0.0271651, 
+0.302108, 0.226044, 0.16552, 0.129912, 0.104309, 0.0754822, 0.0393929, 0.0272213, 0.0271651, 0.0271651, 0.0271651, 
+0.301436, 0.224219, 0.165149, 0.129966, 0.103756, 0.0739283, 0.0370946, 0.0271651, 0.0271651, 0.0271651, 0.0271651,
